chore: refactor: rename varibale and echo output
