+++ src/Views/Form/defaultFormDonationProcessing.php
@@ -4,10 +4,10 @@
  *
  * @since 2.7.0
  */
-use Give\IframeView;
+use Give\Views\IframeView;
 
-$tm = new IframeView();
+$iframeView = new IframeView();
 
-$tm->setTitle( __( 'Donation Processing', 'give' ) )
+echo $iframeView->setTitle( __( 'Donation Processing', 'give' ) )
 	->setBody( apply_filters( 'give_payment_confirm_' . give_clean( $_GET['payment-confirmation'] ), '' ) )
 	->render();
