+++ src/Views/Form/defaultFormReceiptTemplate.php
@@ -5,10 +5,10 @@
  * @since 2.7.0
  */
 use function Give\Helpers\Frontend\getReceiptShortcodeFromConfirmationPage;
-use Give\IframeView;
+use Give\Views\IframeView;
 
-$tm = new IframeView();
+$iframeView = new IframeView();
 
-$tm->setTitle( __( 'Donation Receipt', 'give' ) )
+echo $iframeView->setTitle( __( 'Donation Receipt', 'give' ) )
    ->setBody( do_shortcode( getReceiptShortcodeFromConfirmationPage() ) )
    ->render();
