+++ src/Views/Form/defaultFormTemplate.php
@@ -4,16 +4,16 @@
  *
  * @since 2.7.0
  */
-use Give\IframeView;
+use Give\Views\IframeView;
 use function Give\Helpers\Form\Template\Utils\Frontend\getFormId;
 
-$formId = getFormId();
-$tm     = new IframeView();
+$formId     = getFormId();
+$iframeView = new IframeView();
 
 // Fetch the Give Form.
 ob_start();
 give_get_donation_form( [ 'id' => $formId ] );
 
-$tm->setTitle( get_post_field( 'post_title', $formId ) )
+echo $iframeView->setTitle( get_post_field( 'post_title', $formId ) )
    ->setBody( ob_get_clean() )
    ->render();
