+++ src/Views/Form/defaultRedirectHandlerTemplate.php
@@ -4,7 +4,7 @@
  *
  * @since 2.7.0
  */
-use Give\IframeView;
+use Give\Views\IframeView;
 
 $bodyContent = sprintf(
 	'<p style="text-align: center">%1$s/p>
@@ -17,7 +17,7 @@ $bodyContent = sprintf(
 	esc_js( $location )
 );
 
-$tm = new IframeView();
-$tm->setTitle( __( 'Donation Processing...', 'give' ) )
+$iframeView = new IframeView();
+echo $iframeView->setTitle( __( 'Donation Processing...', 'give' ) )
    ->setBody( $bodyContent )
    ->render();
